// dashboard/redux/tag-types/tag-types.ts
export enum tagTypes {
  user = "users",
  categories = "categories",
  product = "product",
  productSize = "productSize",
  productColor = "productColor",
  orders = "orders",
  payment = "payment",
  qa = "qa",
  tag = "tag",
  testimonial = "testimonial",
  promo = "promo",
  promotionalOffer = "promotionalOffer",
  blogs = "blogs",
  comments = "comments",
  tax = "tax",
  kids = "kids",
}

export const tagTypesList = [
  tagTypes.user,
  tagTypes.categories,
  tagTypes.product,
  tagTypes.productSize,
  tagTypes.productColor,
  tagTypes.orders,
  tagTypes.payment,
  tagTypes.qa,
  tagTypes.tag,
  tagTypes.testimonial,
  tagTypes.promo,
  tagTypes.promotionalOffer,
  tagTypes.blogs,
  tagTypes.comments,
  tagTypes.tax,
  tagTypes.kids,
];